# switch to using @ebay/nodash/get