@yomed 2019-03-12
Hey @scttdavs, thanks for the PR. We've had some discussions around removing lodash here, but the primary reason to do so is actually to avoid lodash's internal cache. On apps with heavy usage, the cache ends up being very large and requires frequent garbage collection, thereby significantly increasing CPU usage.

I wonder if `nodash` caching might have the same issues:
https://github.com/eBay/nodash/blob/7449e927398d0d7c3ec0bb5c0c95bd7a15d01c92/src/getPathArray.js

cc @DylanPiercey