@scttdavs 2019-03-12
if paths are variable (different depending on user), then it probably would. I could maybe add an option to disable cache or even remove it entirely.. But even large apps should not have *that* many different paths right? How large was the cache growing to? Was there a comparison done for performance costs to generate the paths every time?