@yomed 2019-03-12
Some time ago there was investigation done by comparing CPU/GC time for the current lodash version vs. an in-house version without caching, and it found GC improvements. Though the fix wasn't prioritized at the time. I do have the code ready from that fix, so I think I'll put up that PR as well, and then we can use mirroring to test both that PR and this one. That should help validate whether it's the cache that is the issue, or maybe something else.